fix for tests

+++ src/cpu/tests/inc/app.h
@@ -50,7 +50,7 @@ uint8_t get_msg_id(void);
 
 uint16_t fpga_read(uint16_t bram_addr);
 void fpga_write(uint16_t bram_addr, uint16_t value);
-uint16_t bram_read_raw(uint8_t bram_select, size_t bram_addr);
+uint16_t bram_read_raw(uint8_t bram_select, uint32_t bram_addr);
 
 inline static uint16_t get_addr(uint8_t bram_select, uint16_t bram_addr) { return (((uint16_t)bram_select & 0x0003) << 14) | (bram_addr & 0x3FFF); }
